[BACKEND]
[Branch: develop] Improve patch methos at Notification and Activity
Services.

+++ src/backend/fitConnect/fitConnect/src/main/java/com/fitconnet/service/implementations/ActivityServiceImpl.java
@@ -1,11 +1,12 @@
 package com.fitconnet.service.implementations;
 
+import java.security.InvalidParameterException;
 import java.util.Comparator;
 import java.util.Date;
-import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
+import java.util.function.Consumer;
 import java.util.stream.Collectors;
 
 import org.springframework.http.HttpStatus;
@@ -15,6 +16,9 @@ import com.fitconnet.error.exception.activity.ActivityNotFoundException;
 import com.fitconnet.persitence.model.Activity;
 import com.fitconnet.persitence.repository.ActivityRepository;
 import com.fitconnet.service.interfaces.ActivityServiceI;
+import com.fitconnet.utils.Constants;
+
+import jakarta.validation.ConstraintViolationException;
 
 @Service
 public class ActivityServiceImpl implements ActivityServiceI {
@@ -29,7 +33,7 @@ public class ActivityServiceImpl implements ActivityServiceI {
 	@Override
 	public Optional<Set<Activity>> getAll() {
 		List<Activity> activitiesList = activityRepository.findAll();
-		Set<Activity> sortedActivities = new LinkedHashSet();
+		Set<Activity> sortedActivities;
 		sortedActivities = activitiesList.stream().sorted(Comparator.comparing(Activity::getDate))
 				.collect(Collectors.toSet());
 		return Optional.of(sortedActivities);
@@ -38,7 +42,7 @@ public class ActivityServiceImpl implements ActivityServiceI {
 	@Override
 	public Optional<Activity> getOne(Long id) {
 		Activity activity = activityRepository.findById(id)
-				.orElseThrow(() -> new ActivityNotFoundException("Activity not found", HttpStatus.NOT_FOUND));
+				.orElseThrow(() -> new ActivityNotFoundException(Constants.ACTIVITY_NOT_FOUND, HttpStatus.NOT_FOUND));
 		return Optional.of(activity);
 	}
 
@@ -49,8 +53,9 @@ public class ActivityServiceImpl implements ActivityServiceI {
 
 	@Override
 	public void update(Long id, Activity activity) {
-		Activity findActivity = activityRepository.findById(id)
-				.orElseThrow(() -> new ActivityNotFoundException("Activity not found", HttpStatus.NOT_FOUND));
+		if (!activityRepository.findById(id).isPresent()) {
+			throw new ActivityNotFoundException(Constants.ACTIVITY_NOT_FOUND, HttpStatus.NOT_FOUND);
+		}
 
 		activityRepository.save(activity);
 	}
@@ -59,22 +64,18 @@ public class ActivityServiceImpl implements ActivityServiceI {
 	public void patch(Long id, Activity activity) {
 		Activity aux = activityRepository.findById(id)
 				.orElseThrow(() -> new ActivityNotFoundException("Activity not found", HttpStatus.NOT_FOUND));
-		if (aux.getType() != null) {
-			aux.setType(activity.getType());
-		}
-		if (aux.getDuration() != null) {
-			aux.setDuration(activity.getDuration());
-		}
-		if (aux.getPlace() != null) {
-			aux.setPlace(activity.getPlace());
-		}
-
+		updateFieldIfDifferent(aux, activity.getType(), "type", aux::setType);
+		updateFieldIfDifferent(aux, activity.getDuration(), "duration", aux::setDuration);
+		updateFieldIfDifferent(aux, activity.getPlace(), "place", aux::setPlace);
+		updateFieldIfDifferent(aux, activity.getParticipants(), "participants", aux::setParticipants);
+		updateFieldIfDifferent(aux, activity.getDate(), "date", aux::setDate);
 	}
 
 	@Override
 	public void deleteById(Long id) {
-		Activity activity = activityRepository.findById(id)
-				.orElseThrow(() -> new ActivityNotFoundException("Activity not found", HttpStatus.NOT_FOUND));
+		if (!activityRepository.findById(id).isPresent()) {
+			throw new ActivityNotFoundException(Constants.ACTIVITY_NOT_FOUND, HttpStatus.NOT_FOUND);
+		}
 		activityRepository.deleteById(id);
 
 	}
@@ -98,6 +99,34 @@ public class ActivityServiceImpl implements ActivityServiceI {
 		newActivity.setPlace(activity.getPlace());
 		newActivity.setType(activity.getType());
 
+	}
+
+	private <T> void updateFieldIfDifferent(Activity activity, T newValue, String fieldName, Consumer<T> setter) {
+		T existingValue = getFieldValue(activity, fieldName);
+		if (newValue != null && !newValue.equals(existingValue)) {
+			try {
+				setter.accept(newValue);
+			} catch (ConstraintViolationException e) {
+				throw new InvalidParameterException("El valor para '" + fieldName + "' no es válido.");
+			}
+		}
+	}
+
+	private <T> T getFieldValue(Activity activity, String fieldName) {
+		switch (fieldName) {
+		case "type":
+			return (T) activity.getType();
+		case "duration":
+			return (T) activity.getDuration();
+		case "place":
+			return (T) activity.getPlace();
+		case "participants":
+			return (T) activity.getParticipants();
+		case "date":
+			return (T) activity.getDate();
+		default:
+			throw new IllegalArgumentException("Campo desconocido: " + fieldName);
+		}
 	}
 
 }
